requirements.txt
streamlit
pillow
streamlit_drawable_canvas
google-cloud-aiplatform